Clean up and write doc-comment for CodeWriter::writeNip().

+++ core/Verifier.cpp
@@ -2823,7 +2823,7 @@ namespace avmplus
             FrameValue v = state->stackTop();
             // NOTE writeNip is necessary until lir optimizes the "nip"
             // case to avoid the extra copies that result from swap+pop
-            coder->writeNip(state, pc, (uint8_t) argc);
+            coder->writeNip(state, pc, argc);
             state->pop(argc + 1);
             state->push(v);
         }
